implement entity level authorization

+++ apps/gateway/src/core/application/services/scheduled-task/delete-scheduled-task.service.ts
@@ -1,4 +1,10 @@
-import { Inject, Logger, OnModuleDestroy, OnModuleInit } from '@nestjs/common';
+import {
+  Inject,
+  Logger,
+  OnModuleDestroy,
+  OnModuleInit,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
 import { DeleteScheduledTaskCommand } from 'apps/gateway/src/interface/commands/scheduled-task/delete-scheduled-task.command';
 import { AggregateID } from 'libs/ddd/entity.base';
@@ -8,8 +14,6 @@ import { ClientProxy } from '@nestjs/microservices';
 import { TransportAdapterNames } from 'libs/common/enum/adapters/adapters.enum';
 import { firstValueFrom } from 'rxjs';
 import { ScheduledTaskIntegrationEvents } from 'libs/events/scheduled-task.events';
-
-// TODO... only delete scheduled task the user is an owner of
 
 @CommandHandler(DeleteScheduledTaskCommand)
 export class DeleteScheduledTaskService
@@ -32,6 +36,11 @@ export class DeleteScheduledTaskService
     try {
       await this.repo.transaction(async () => {
         const task = await this.repo.findOneById(command.id);
+        if (!task.isOwner(command.userId)) {
+          throw new UnauthorizedException(
+            'User does not have permission to delete resource',
+          );
+        }
         task.delete();
         await this.repo.delete(task);
         await firstValueFrom(
